# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A simple RESTful API server built with Go, Gin, sqlc, and golang-migrate.",
        "title": "Mini Lab API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.minilab.com/support",
            "email": "support@minilab.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/machine": {
            "get": {
                "description": "Get a list of all machines",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Get all machines",
                "responses": {
                    "200": {
                        "description": "List of machines",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Machine"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new machine with optional task type assignment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Create a new machine",
                "parameters": [
                    {
                        "description": "Machine creation data",
                        "name": "machine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMachineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Machine created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Machine"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/machine/{id}": {
            "get": {
                "description": "Get a single machine by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Get machine by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Machine"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid machine ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Machine not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing machine by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Update a machine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Machine update data",
                        "name": "machine",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMachineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Machine"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Machine not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a machine by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Delete a machine",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Machine deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid machine ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Machine not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "Error message"
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.CreateMachineRequest": {
            "type": "object",
            "required": [
                "estimate_time",
                "name",
                "quantity"
            ],
            "properties": {
                "estimate_time": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 120
                },
                "name": {
                    "type": "string",
                    "example": "3D Printer Model X"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                },
                "task_type": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.Machine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "estimate_time": {
                    "description": "in minutes",
                    "type": "integer",
                    "example": 120
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "3D Printer Model X"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "task_type": {
                    "type": "string",
                    "example": "3D Printing"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
        "models.UpdateMachineRequest": {
            "type": "object",
            "properties": {
                "estimate_time": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 90
                },
                "name": {
                    "type": "string",
                    "example": "3D Printer Model Y"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 3
                }
            }
        },
...
    }
}